// src/types/swagger_output_param_info.ts
import { SwaggerRef } from "./swagger_ref";
import { SwaggerCustomParam } from "./swagger_custom_param";
import { SWAGGER_OUTPUT_PARAM } from "../enums/swagger_output_param";

export type SwaggerOutputParamInfo = {
    name: string;
    in: SWAGGER_OUTPUT_PARAM,
    description?: string;
    required: boolean;
    schema: SwaggerRef | SwaggerCustomParam
}